# [Proposal] Accept a Set for metric measurement operations