--- a/exporters/prometheus/benchmark_test.go
+++ b/exporters/prometheus/benchmark_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/require"
 
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/sdk/metric"
 )
 
@@ -36,7 +37,7 @@ func benchmarkCollect(b *testing.B, n int) {
 	for i := 0; i < n; i++ {
 		counter, err := meter.Float64Counter(fmt.Sprintf("foo_%d", i))
 		require.NoError(b, err)
-		counter.Add(ctx, float64(i))
+		counter.Add(ctx, float64(i), *attribute.EmptySet())
 	}
 
 	b.ReportAllocs()