--- a/metric/example_test.go
+++ b/metric/example_test.go
@@ -40,7 +40,7 @@ func ExampleMeter_synchronous() {
 	ctx := context.Background()
 	// Do work
 	// ...
-	workDuration.Record(ctx, time.Since(startTime).Milliseconds())
+	workDuration.Record(ctx, time.Since(startTime).Milliseconds(), *attribute.EmptySet())
 }
 
 func ExampleMeter_asynchronous_single() {
@@ -63,7 +63,8 @@ func ExampleMeter_asynchronous_single() {
 			//
 			// For demonstration purpose, a static value is used here.
 			usage := 75000
-			obsrv.Observe(int64(usage), attribute.Int("disk.id", 3))
+			attrs := attribute.NewSet(attribute.Int("disk.id", 3))
+			obsrv.Observe(int64(usage), attrs)
 			return nil
 		}),
 	)
@@ -87,8 +88,8 @@ func ExampleMeter_asynchronous_multiple() {
 			// This call does work
 			runtime.ReadMemStats(memStats)
 
-			o.ObserveInt64(heapAlloc, int64(memStats.HeapAlloc))
-			o.ObserveInt64(gcCount, int64(memStats.NumGC))
+			o.ObserveInt64(heapAlloc, int64(memStats.HeapAlloc), *attribute.EmptySet())
+			o.ObserveInt64(gcCount, int64(memStats.NumGC), *attribute.EmptySet())
 
 			// This function synchronously records the pauses
 			computeGCPauses(ctx, gcPause, memStats.PauseNs[:])