--- a/metric/instrument/asyncfloat64.go
+++ b/metric/instrument/asyncfloat64.go
@@ -181,7 +181,7 @@ type Float64ObservableGaugeOption interface {
 // Warning: methods may be added to this interface in minor releases.
 type Float64Observer interface {
 	// Observe records the float64 value with attributes.
-	Observe(value float64, attributes ...attribute.KeyValue)
+	Observe(value float64, attributes attribute.Set)
 }
 
 // Float64Callback is a function registered with a Meter that makes