--- a/metric/instrument/asyncfloat64_test.go
+++ b/metric/instrument/asyncfloat64_test.go
@@ -47,7 +47,7 @@ func TestFloat64ObservableConfiguration(t *testing.T) {
 	}
 
 	cback := func(ctx context.Context, obsrv Float64Observer) error {
-		obsrv.Observe(token)
+		obsrv.Observe(token, *attribute.EmptySet())
 		return nil
 	}
 
@@ -87,6 +87,6 @@ type float64Observer struct {
 	got float64
 }
 
-func (o *float64Observer) Observe(v float64, _ ...attribute.KeyValue) {
+func (o *float64Observer) Observe(v float64, _ attribute.Set) {
 	o.got = v
 }