--- a/metric/instrument/asyncint64.go
+++ b/metric/instrument/asyncint64.go
@@ -180,7 +180,7 @@ type Int64ObservableGaugeOption interface {
 // Warning: methods may be added to this interface in minor releases.
 type Int64Observer interface {
 	// Observe records the int64 value with attributes.
-	Observe(value int64, attributes ...attribute.KeyValue)
+	Observe(value int64, attributes attribute.Set)
 }
 
 // Int64Callback is a function registered with a Meter that makes observations