--- a/metric/meter.go
+++ b/metric/meter.go
@@ -127,9 +127,9 @@ type Callback func(context.Context, Observer) error
 // Observer records measurements for multiple instruments in a Callback.
 type Observer interface {
 	// ObserveFloat64 records the float64 value with attributes for obsrv.
-	ObserveFloat64(obsrv instrument.Float64Observable, value float64, attributes ...attribute.KeyValue)
+	ObserveFloat64(obsrv instrument.Float64Observable, value float64, attributes attribute.Set)
 	// ObserveInt64 records the int64 value with attributes for obsrv.
-	ObserveInt64(obsrv instrument.Int64Observable, value int64, attributes ...attribute.KeyValue)
+	ObserveInt64(obsrv instrument.Int64Observable, value int64, attributes attribute.Set)
 }
 
 // Registration is an token representing the unique registration of a callback