--- a/sdk/metric/instrument.go
+++ b/sdk/metric/instrument.go
@@ -180,24 +180,23 @@ var _ instrument.Int64Counter = (*instrumentImpl[int64])(nil)
 var _ instrument.Int64UpDownCounter = (*instrumentImpl[int64])(nil)
 var _ instrument.Int64Histogram = (*instrumentImpl[int64])(nil)
 
-func (i *instrumentImpl[N]) Add(ctx context.Context, val N, attrs ...attribute.KeyValue) {
+func (i *instrumentImpl[N]) Add(ctx context.Context, val N, attrs attribute.Set) {
 	i.aggregate(ctx, val, attrs)
 }
 
-func (i *instrumentImpl[N]) Record(ctx context.Context, val N, attrs ...attribute.KeyValue) {
+func (i *instrumentImpl[N]) Record(ctx context.Context, val N, attrs attribute.Set) {
 	i.aggregate(ctx, val, attrs)
 }
 
-func (i *instrumentImpl[N]) aggregate(ctx context.Context, val N, attrs []attribute.KeyValue) {
+func (i *instrumentImpl[N]) aggregate(ctx context.Context, val N, attrs attribute.Set) {
 	if err := ctx.Err(); err != nil {
 		return
 	}
 	// Do not use single attribute.Sortable and attribute.NewSetWithSortable,
 	// this method needs to be concurrent safe. Let the sync.Pool in the
 	// attribute package handle allocations of the Sortable.
-	s := attribute.NewSet(attrs...)
 	for _, agg := range i.aggregators {
-		agg.Aggregate(val, s)
+		agg.Aggregate(val, attrs)
 	}
 }
 
@@ -261,13 +260,12 @@ func newObservable[N int64 | float64](scope instrumentation.Scope, kind Instrume
 }
 
 // observe records the val for the set of attrs.
-func (o *observable[N]) observe(val N, attrs []attribute.KeyValue) {
+func (o *observable[N]) observe(val N, attrs attribute.Set) {
 	// Do not use single attribute.Sortable and attribute.NewSetWithSortable,
 	// this method needs to be concurrent safe. Let the sync.Pool in the
 	// attribute package handle allocations of the Sortable.
-	s := attribute.NewSet(attrs...)
 	for _, agg := range o.aggregators {
-		agg.Aggregate(val, s)
+		agg.Aggregate(val, attrs)
 	}
 }