--- a/sdk/metric/meter.go
+++ b/sdk/metric/meter.go
@@ -292,7 +292,7 @@ var (
 	errUnregObserver   = errors.New("observable instrument not registered for callback")
 )
 
-func (r observer) ObserveFloat64(o instrument.Float64Observable, v float64, a ...attribute.KeyValue) {
+func (r observer) ObserveFloat64(o instrument.Float64Observable, v float64, a attribute.Set) {
 	var oImpl float64Observable
 	switch conv := o.(type) {
 	case float64Observable:
@@ -324,7 +324,7 @@ func (r observer) ObserveFloat64(o instrument.Float64Observable, v float64, a ..
 	oImpl.observe(v, a)
 }
 
-func (r observer) ObserveInt64(o instrument.Int64Observable, v int64, a ...attribute.KeyValue) {
+func (r observer) ObserveInt64(o instrument.Int64Observable, v int64, a attribute.Set) {
 	var oImpl int64Observable
 	switch conv := o.(type) {
 	case int64Observable:
@@ -417,7 +417,7 @@ type int64Observer struct {
 	int64Observable
 }
 
-func (o int64Observer) Observe(val int64, attrs ...attribute.KeyValue) {
+func (o int64Observer) Observe(val int64, attrs attribute.Set) {
 	o.observe(val, attrs)
 }
 
@@ -448,6 +448,6 @@ type float64Observer struct {
 	float64Observable
 }
 
-func (o float64Observer) Observe(val float64, attrs ...attribute.KeyValue) {
+func (o float64Observer) Observe(val float64, attrs attribute.Set) {
 	o.observe(val, attrs)
 }