Resolves https://github.com/open-telemetry/opentelemetry-go/issues/3943

- Prevents data races when a user passes the same attribute slice to multiple measurement operations concurrently. All measurement methods are required by the specification to concurrent safe[^1].
- Allows users to allocate a single `attribute.Set` for all measurement methods instead of having each method create its own `attribute.Set` for the same data.

[^1]: https://github.com/open-telemetry/opentelemetry-specification/blob/9352c2524ac03b31f4b845323b351d32ab3adb43/specification/metrics/api.md#concurrency-requirements