@MrAlias 2023-03-30
### Benchmark of counter Add

- `old.txt`: f4a9d78e
- `new.txt`: 3425a1a0

```terminal
$ benchstat old.txt new.txt
name                                old time/op    new time/op    delta
CounterAddNoAttrs-8                   42.0ns ± 6%    42.2ns ± 6%      ~     (p=0.529 n=10+10)
CounterAddOneAttr-8                    258ns ± 8%     104ns ± 6%   -59.66%  (p=0.000 n=10+10)
CounterAddOneInvalidAttr-8             361ns ± 9%     172ns ± 7%   -52.45%  (p=0.000 n=10+10)
CounterAddSingleUseAttrs-8             851ns ± 1%     843ns ± 3%      ~     (p=0.156 n=9+10)
CounterAddSingleUseInvalidAttrs-8     1.24µs ± 3%    1.22µs ± 1%    -1.67%  (p=0.008 n=10+10)
CounterAddSingleUseFilteredAttrs-8    1.38µs ± 3%    1.36µs ± 1%    -1.37%  (p=0.011 n=10+10)
CounterCollectOneAttr-8                300ns ± 8%     145ns ± 2%   -51.76%  (p=0.000 n=10+10)
CounterCollectTenAttrs-8              2.87µs ± 6%    2.86µs ±10%      ~     (p=0.617 n=10+10)

name                                old alloc/op   new alloc/op   delta
CounterAddNoAttrs-8                    0.00B          0.00B           ~     (all equal)
CounterAddOneAttr-8                     128B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
CounterAddOneInvalidAttr-8              256B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
CounterAddSingleUseAttrs-8              226B ± 0%      225B ± 1%      ~     (p=0.054 n=8+9)
CounterAddSingleUseInvalidAttrs-8       379B ± 0%      379B ± 0%      ~     (all equal)
CounterAddSingleUseFilteredAttrs-8      571B ± 0%      571B ± 0%      ~     (all equal)
CounterCollectOneAttr-8                 144B ± 0%       16B ± 0%   -88.89%  (p=0.000 n=10+10)
CounterCollectTenAttrs-8              1.30kB ± 0%    1.30kB ± 0%      ~     (all equal)

name                                old allocs/op  new allocs/op  delta
CounterAddNoAttrs-8                     0.00           0.00           ~     (all equal)
CounterAddOneAttr-8                     2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
CounterAddOneInvalidAttr-8              2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
CounterAddSingleUseAttrs-8              2.00 ± 0%      2.00 ± 0%      ~     (all equal)
CounterAddSingleUseInvalidAttrs-8       2.00 ± 0%      2.00 ± 0%      ~     (all equal)
CounterAddSingleUseFilteredAttrs-8      4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CounterCollectOneAttr-8                 3.00 ± 0%      1.00 ± 0%   -66.67%  (p=0.000 n=10+10)
CounterCollectTenAttrs-8                21.0 ± 0%      21.0 ± 0%      ~     (all equal)
```