@merlimat 2023-03-30
@MrAlias Nice comparison! It would be great to also show the difference when one can keep reusing the attributes set.

```go
// Old API
func BenchmarkCounterAddMultiUseAttrs(b *testing.B) {
	ctx, _, cntr := benchCounter(b)

       attrs := []attribute.KeyValue{attribute.Int("K", i)}

	for i := 0; i < b.N; i++ {
		cntr.Add(ctx, 1, attrs...)
	}
}
```

```go
// New API
func BenchmarkCounterAddMultiUseAttrs(b *testing.B) {
	ctx, _, cntr := benchCounter(b)

       attrs := attribute.NewSet(attribute.Int("K", i))

	for i := 0; i < b.N; i++ {
		cntr.Add(ctx, 1, attrs)
	}
}
```