@merlimat 2023-03-31
thanks! My bad, I couldn't find the updated benchmark code in the PR.