@MrAlias 2023-04-07
Closing in favor of https://github.com/open-telemetry/opentelemetry-go/pull/3971